v3-030 mirror vendor negative-immediate quirk in source c-model
The vendor executable reinterprets sign-extended int16 immediates as
uint32, turning 0xFFFD (-3) into ~4.29e9 (add -> inf, x0 -> 0, div -> 0)
in the elementwise and broadcast immediate paths; the shift amount stays
signed. Reverse-engineered on a.out, mirrored in mysim_0818.cpp per the
bug-parity policy, locked with test_negative_immediate_parity. Compiler
output is unaffected (only imm=0 is ever emitted; real constants travel
as FP16 memory). Recorded as V3-030 and ISA review 7.3-9.

+++ _poc/mysim_0818.cpp
@@ -323,7 +323,12 @@ private:
         if (mode == 1) {
             return read_g(immediate);
         }
-        return static_cast<float>(static_cast<std::int16_t>(immediate));
+        // Vendor quirk: the sign-extended immediate is reinterpreted as uint32,
+        // so a negative immediate becomes ~4.29e9 (0xFFFD -> 4294967293.0f).
+        // Only 0..32767 behave as written.  (Verified against a.out; the shift
+        // amount path is NOT affected and stays signed.)
+        return static_cast<float>(static_cast<std::uint32_t>(
+            static_cast<std::int16_t>(immediate)));
     }
 
     void matmul(std::uint32_t instruction, unsigned activation) {
@@ -480,7 +485,10 @@ private:
                 broadcast_address_ = replace_half(broadcast_address_, value, high);
                 scalar = read_g(broadcast_address_);
             } else {
-                scalar = static_cast<float>(static_cast<std::int16_t>(value));
+                // Same uint32 reinterpretation quirk as the elementwise
+                // immediate path (negative broadcast immediates become huge).
+                scalar = static_cast<float>(static_cast<std::uint32_t>(
+                    static_cast<std::int16_t>(value)));
             }
             output_.assign(vector_length_, scalar);
             output_rows_ = 1;
